requirements.txt
alembic==1.7.5
Flask==1.1.4
Flask_Bcrypt==0.7.1
Flask_JWT_Extended==3.18.0
Flask_Migrate==3.1.0
Flask_OAuthlib==0.9.6
flask_restx==0.5.1
Flask_Script==2.0.5
Flask_SQLAlchemy==2.5.1
flask_testing==0.8.1
protobuf==4.21.12
pygit2==1.9.1
PyJWT==2.6.0
python_telegram_bot==13.8.1
python-dotenv==0.21.1
pylint==2.15.9
psycopg2==2.9.5
SQLAlchemy==1.4.27
SQLAlchemy_Utils==0.38.2
telegram==0.0.1
waitress==2.0.0
markupsafe==2.0.1
flake8==6.0.0
